refactor: split `BigNat` into two classes
